+++ apps/block_scout_web/lib/block_scout_web/views/api/v2/transaction_view.ex
@@ -368,7 +368,7 @@ defmodule BlockScoutWeb.API.V2.TransactionView do
       "result" => status,
       "status" => transaction.status,
       "block" => transaction.block_number,
-      "timestamp" => block_timestamp(transaction.block),
+      "timestamp" => block_timestamp(transaction),
       "from" =>
         Helper.address_with_info(
           single_tx? && conn,
@@ -799,6 +799,7 @@ defmodule BlockScoutWeb.API.V2.TransactionView do
     end
   end
 
+  defp block_timestamp(%Transaction{block_timestamp: block_ts}) when not is_nil(block_ts), do: block_ts
   defp block_timestamp(%Transaction{block: %Block{} = block}), do: block.timestamp
   defp block_timestamp(%Block{} = block), do: block.timestamp
   defp block_timestamp(_), do: nil